+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2023-12-26
+## [Unreleased] - 2023-12-27
 ### Added
 - New `viteclientts` folder for migrating `viteclient` from JavaScript to TypeScript.
 - Copied `images` folder to the new `viteclientts` folder.
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migrated `index`, `ScheduledTaskCard` and `SchedulerCard` components from JavaScript to TypeScript.
 - Modified Types of setSelectedHour and setSelectedMinute to strings
 - Migrated `Layout` and `ErrorBoundary` from JavaScript to TypeScript.
+- Migrated `404Page`, `Homepage`, `LoginPage` and `SettingsPage` from JavaScript to TypeScript.
+- Replaced react-cookie with universal-cookie to avoid Typescript error in Loginpage
 
 ## [v15.7.2] - 2023-12-25
 ### Changed